# bpo-42834: Make static cache variables in _json.c compatible with subinterpreters