--- a/Misc/NEWS.d/next/Library/2021-01-05-23-55-24.bpo-42834.LxRnZC.rst
+++ b/Misc/NEWS.d/next/Library/2021-01-05-23-55-24.bpo-42834.LxRnZC.rst
@@ -1,0 +1,2 @@
+Convert internal caches in the ``_json.c`` module from static variables to
+heap variables.