--- a/Misc/NEWS.d/next/Library/2021-01-05-23-55-24.bpo-42834.LxRnZC.rst
+++ b/Misc/NEWS.d/next/Library/2021-01-05-23-55-24.bpo-42834.LxRnZC.rst
@@ -1,2 +1,2 @@
-Convert internal caches in the ``_json.c`` module from static variables to
-heap variables.
+Convert static internal caches in the ``_json.c`` module to work with
+subinterpreters.