--- a/Misc/NEWS.d/next/Library/2021-01-05-23-55-24.bpo-42834.LxRnZC.rst
+++ b/Misc/NEWS.d/next/Library/2021-01-05-23-55-24.bpo-42834.LxRnZC.rst
@@ -1,2 +1,1 @@
-Convert static internal caches in the ``_json.c`` module to work with
-subinterpreters.
+Make internal caches of the ``_json`` module compatible with subinterpreters.