This patch helps isolate static caches in _json.c, and *may* help with decreasing refleaks at finalization too.

No measured performance slowdowns in json_loads and json_dumps in pyperformance with this patch. There is a measurable slowdown in error raising of around 8%, but this isn't usually very hot code.

<!-- issue-number: [bpo-42834](https://bugs.python.org/issue42834) -->
https://bugs.python.org/issue42834
<!-- /issue-number -->