@Fidget-Spinner 2021-01-06
> Did you consider to use _Py_IDENTIFIER() API? It is compatible with subinterpreters and it provides interned strings. _PyUnicode_FromId() can fail with memory allocation error (at the first call, when the string is created). There are "Id" variant of many functions.

No that didn't cross my mind due to my inexperience with the internal C API. Thank you for taking the time to review this! Please give me a week to address your suggestions as I'm currently quite busy. Thanks!