@Fidget-Spinner 2021-01-10
More benchmarking results:
main branch -> current implementation (using state to store) -> using _Py_IDENTIFIER , static error -> using _Py_IDENTIFIER, importing error every time
```
json_dumps
==========

Mean +- std dev: [json-main] 19.4 ms +- 1.2 ms -> [json-heap] 21.8 ms +- 1.4 ms: 1.12x slower (+12%)
Mean +- std dev: [json-main] 19.4 ms +- 1.2 ms -> [json-pyid] 21.3 ms +- 2.6 ms: 1.09x slower (+9%)
Mean +- std dev: [json-main] 19.4 ms +- 1.2 ms -> [json-pyid-noerr] 22.2 ms +- 1.1 ms: 1.14x slower (+14%)

json_loads
==========

Mean +- std dev: [json-main] 46.5 us +- 1.2 us -> [json-heap] 52.8 us +- 1.9 us: 1.14x slower (+14%)
Mean +- std dev: [json-main] 46.5 us +- 1.2 us -> [json-pyid] 49.4 us +- 3.4 us: 1.06x slower (+6%)
Mean +- std dev: [json-main] 46.5 us +- 1.2 us -> [json-pyid-noerr] 52.4 us +- 2.0 us: 1.13x slower (+13%)
```
Using _Py_IDENTIFIER , static error seems best, so I'm going to do that.