@Fidget-Spinner 2021-01-10
Sorry, it seems that I was wrong, my personal computer isn't producing consistent pyperformance results, so I don't see any real slowdown/speedup.