@Fidget-Spinner 2021-01-10
@vstinner I think this should be ready now. Pyperformance on windows (this isn't really accurate, but it's something at least) says:
```
json_dumps: Mean +- std dev: [json-main2] 17.5 ms +- 0.2 ms -> [json-heap] 17.8 ms +- 0.1 ms: 1.01x slower (+1%)
Benchmark hidden because not significant (1): json_loads
```
If it gets merged, I'll monitor speed.python.org for awhile and see if anything crops up. Thanks for your patience and help in all of this!