@Fidget-Spinner 2021-01-29
@vstinner I was finally able to test it on a linux machine (btw, json_dumps is slightly unstable on my machine and pyperformance sometimes complains, only json_loads is stable):
on master branch:
```
### json_dumps ###
Mean +- std dev: 47.5 ms +- 5.6 ms

### json_loads ###
Mean +- std dev: 81.1 us +- 10.5 us
```
on this branch:
```
### json_dumps ###
Mean +- std dev: 48.2 ms +- 4.4 ms

### json_loads ###
Mean +- std dev: 80.0 us +- 8.7 us
```
```
> pyperf compare_to master2.json json-heap2.json 
Benchmark hidden because not significant (2): json_dumps, json_loads

Geometric mean: 1.00x slower
```

Seems like performance isn't affected :).