@vstinner 2021-01-29
> So raising error is quite a common operation. Using PyId here surprisingly seems to speed up error raising by around 20%.

What is your benchmark showing a 20% difference?