@Fidget-Spinner 2021-01-30
> > So raising error is quite a common operation. Using PyId here surprisingly seems to speed up error raising by around 20%.
> 
> What is your benchmark showing a 20% difference?

It's a microbenchmark, so I don't see it as 100% representative of a real-world scenario:
```bash
pyperf timeit "import json" "try: json.loads('{dkfjdkjf')
except: pass" 
```
```
# without static strings:

Mean +- std dev: 9.89 us +- 0.45 us

# with static strings:

Mean +- std dev: 8.36 us +- 0.54 us
```
compare_to
```
Mean +- std dev: [json-heap-no-static] 9.89 us +- 0.45 us -> [json-heap-static-string] 8.36 us +- 0.54 us: 1.18x faster
```