@vstinner 2021-02-01
Can you please compare using the current master branch as a reference?

* bench 1: reference, master
* bench 2: your PR without static strings (expected to be slower)
* bench 3: your PR using static strings

> pyperf timeit "import json" "try: json.loads('{dkfjdkjf')

It seems like you import json at each iteration. Use `-s "import json"`.