@Fidget-Spinner 2021-02-01
> It seems like you import json at each iteration. Use `-s "import json"`.

Wow! I didn't know that. I assumed that the first argument is automatically considered setup code without having to specify ``-s``. Thanks for correcting me.



> Can you please compare using the current master branch as a reference?
> 
>     * bench 1: reference, master
> 
>     * bench 2: your PR without static strings (expected to be slower)
> 
>     * bench 3: your PR using static strings
> 
I did:
- [X] ``sudo pyperf system tune``
- [X] Ensured that I only saw 1us difference between runs before starting the benchmark.

``python3 -m pyperf timeit -s "import json" "try: json.loads('{')
except: pass"``

```
bench 1: reference, master
Mean +- std dev: 8.00 us +- 0.19 us

bench 2: PR without static strings (expected to be slower)
Mean +- std dev: 10.3 us +- 0.2 us

bench 3: PR using static strings
Mean +- std dev: 8.52 us +- 0.26 us

# compare_to bench1, bench2, bench3
python3 -m pyperf compare_to master.json  json-heap-no-pyid.json json-heap-pyid.json 
Mean +- std dev: [master] 8.00 us +- 0.19 us -> [json-heap-no-pyid] 10.3 us +- 0.2 us: 1.28x slower
Mean +- std dev: [master] 8.00 us +- 0.19 us -> [json-heap-pyid] 8.52 us +- 0.26 us: 1.07x slower
```