@vstinner 2021-02-01
Merged, thanks. Well, since the code already uses static strings for the "hot code" (true/false), it's ok to "optimize" the error path (raise an exception).