@Fidget-Spinner 2021-02-01
Thanks for your patience and guidance on this PR Victor! I have waaaay more respect for the all the pyperf/pyperformance people now - benchmarks are pretty hard 😆.