@vstinner 2021-02-01
>  Ensured that I only saw 0.1us difference between runs before starting the benchmark.

I don't understand what you mean. If you don't use CPU isolation, you should run benchmarks on an idle machine (don't run other programs in parallel).

You should not pick which numbers look better to you, but accumulate more data. Use --append option rather than -o/--output when you create JSON files. Using pyperf, you can run a benchmark many times and accumulate more and more runs, pyperf computes the mean and std dev. Then you can have fun with stats and hist commands ;-)

https://pyperf.readthedocs.io/en/latest/analyze.html