@Fidget-Spinner 2021-02-01
> You should not pick which numbers look better to you, but accumulate more data. Use --append option rather than -o/--output when you create JSON files. Using pyperf, you can run a benchmark many times and accumulate more and more runs, pyperf computes the mean and std dev. Then you can have fun with stats and hist commands ;-)
> 
> https://pyperf.readthedocs.io/en/latest/analyze.html

Thanks, TIL something new again about append.

>I don't understand what you mean. If you don't use CPU isolation, you should run benchmarks on an idle machine (don't run other programs in parallel).


Sorry I wasn't clear here. What I meant was: I did that to make sure that all the tuning *was* working, not cherry picking data on purpose. Because on Windows, when I ran pyperf, there was actually still quite a lot of noise between each run (because tuning didn't really work!). So this was just to mean: my machine was quite stable.

On the linux machine I didn't have any other programs running (well other than the desktop environment and terminal :p).