@Fidget-Spinner 2021-02-01
Point taken about accumulating data though - I never knew such a useful feature existed. If I had known I'd definitely use it way more. Thanks!