@vstinner 2021-02-01
Oh, I only run benchmarks on Linux. Someone should write a documentation explaining how to run reproducible benchmarks on Windows. pyperf changes the process priority if it has the required dependency and the permissions. But I don't recall the details.